workaround for ARM compatibility
confluentinc/ksqldb-server:0.28.2 does not support ARM, and throws some strange errors when running via x86 emulation. The cp-ksqldb-server:7.3.2 (which is based on the same ksqldb version) does not have this issue & is compatible with both ARM and x86. But, it keeps its run command in a different directory, so it's necessary to modify the command in the docker-compose.yml to account for this.

Also bumped connector versions to install while I was at it.